+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v0.3.0] - 2022-06-09
+### Added
+- The certificate for the ingress-nginx will be automatically updated.
+A watch recognizes changes in the registry for the certificate and updates the ssl secret [#5].
+
 ## [v0.2.0] - 2022-06-08
 ### Added
 - Warp menu generation